// src/const/type/Station.enum.ts
export enum Station {
    byleth_station = 'byleth_station',
    chrom_station = 'chrom_station',
    corrin_station = 'corrin_station',
    ike_station = 'ike_station',
    lucina_station = 'lucina_station',
    marth_station = 'marth_station',
    robin_station = 'robin_station',
    roy_station = 'roy_station',
}